console: Remove duplicate button
