--- pkg/webui/components/button-v2/button.styl
@@ -1,144 +1,0 @@
-// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.button
-  reset-button()
-  position: relative
-  display: inline-flex
-  transition: 0.2s all ease-in-out
-  border-radius: $br.l
-  outline: 0
-  cursor: pointer
-  align-items: center
-  gap: $cs.xxs
-  height: 2.5rem
-  text-decoration: none
-  padding: 0 $cs.m
-  white-space: nowrap
-
-  .arrow-icon
-    transform: rotate(0deg)
-    transition: transform .2s
-
-  .arrow-icon-expanded
-    transform: rotate(180deg)
-    transition: transform .2s
-
-  &.primary
-    transition: 0.2s all ease-in-out
-    color: white
-    background-color: $c-active-blue
-
-    &.disabled
-      background-color: $c-divider-dark
-
-    &:not(.disabled)
-      &:hover
-        transition: 0.2s all ease-in-out
-        background-color: hoverize($c-active-blue)
-
-      +focus-visible()
-        background-color: hoverize($c-active-blue)
-
-      &:active
-        transition: 0.2s all ease-in-out
-        background-color: activize($c-active-blue)
-
-    &.grey
-      background-color: $c.grey-500
-
-      &:not(.disabled)
-        &:hover
-          background-color: hoverize($c.grey-500)
-
-        +focus-visible()
-          background-color: hoverize($c.grey-500)
-
-        &:active
-          background-color: activize($c.grey-500)
-
-      &.disabled
-        background-color: transparentify($c.grey-500, #fff, .25)
-
-  &.secondary
-    transition: 0.2s all ease-in-out
-    color: $tc-deep-gray
-    background-color: white
-    box-shadow: 0px 2px 2px 0px rgba(0 0 0 4.5%), inset 0px 0px 0px 1px white
-    background: linear-gradient(to bottom, white, white) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
-    border: 1px solid transparent
-
-
-    &:hover
-      transition: 0.2s all ease-in-out
-      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
-
-    +focus-visible()
-      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
-
-    &:active
-      transition: 0.2s all ease-in-out
-      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
-      box-shadow: none
-
-  &.naked
-    transition: 0.2s all ease-in-out
-    border: none
-    background-color: none
-    box-shadow: unset
-    color: $tc-subtle-gray
-
-    &:not(.disabled)
-      &:hover
-        background-color: $c-backdrop
-
-      +focus-visible()
-        background-color: $c-backdrop
-
-      &:active
-        transition-duration: $ad.xs
-        background-color: $c-backdrop
-
-  &.only-icon
-    width: 2.5rem
-    padding: 0 $cs.xs 0 0.73rem
-
-  &.with-dropdown
-    padding: 0 $cs.xxs 0 $cs.xs
-
-  &.with-dropdown&.only-icon
-    width: fit-content
-    .icon
-      margin-left: 0
-
-  &.with-icon
-    line-height: 1.3
-
-    &.responsive-label
-      padding: $cs.xs
-
-      +media-query($bp.xxs)
-        .icon
-          margin-left: 0
-          margin-right: 0
-
-        .link-button-message
-          display: none
-
-.content
-  position: relative
-
-.icon
-  margin-left: - $cs.xxs
-
