+++ pkg/webui/components/button/button.styl
@@ -1,4 +1,4 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,38 +14,63 @@
 
 .button
   reset-button()
+  position: relative
   display: inline-flex
-  transition: color $ad.s, background $ad.s, border $ad.s, opacity $ad.s
-  border-radius: $br.xs
+  transition: 0.2s all ease-in-out
+  border-radius: $br.l
   outline: 0
   cursor: pointer
   justify-content: center
   align-items: center
-  height: $default-input-height
-  font-size: 1rem
+  gap: $cs.xxs
+  height: 2.5rem
   text-decoration: none
-  padding: 0 $cs.s
+  padding: 0 $cs.m
   white-space: nowrap
 
+  .arrow-icon
+    transform: rotate(0deg)
+    transition: transform .2s
+
+  .arrow-icon-expanded
+    transform: rotate(180deg)
+    transition: transform .2s
+
   &.primary
+    transition: 0.2s all ease-in-out
     color: white
     background-color: $c-active-blue
-    padding: 0 $cs.s
-    font-weight: $fw.bold
 
     &.disabled
       background-color: $c-divider-dark
 
     &:not(.disabled)
       &:hover
+        transition: 0.2s all ease-in-out
         background-color: hoverize($c-active-blue)
 
       +focus-visible()
         background-color: hoverize($c-active-blue)
 
       &:active
-        transition-duration: $ad.xs
+        transition: 0.2s all ease-in-out
         background-color: activize($c-active-blue)
+
+    &.grey
+      background-color: $c.grey-500
+
+      &:not(.disabled)
+        &:hover
+          background-color: hoverize($c.grey-500)
+
+        +focus-visible()
+          background-color: hoverize($c.grey-500)
+
+        &:active
+          background-color: activize($c.grey-500)
+
+      &.disabled
+        background-color: transparentify($c.grey-500, #fff, .25)
 
     &.warning
       background-color: $c-warning
@@ -76,11 +101,32 @@
       &.disabled
         background-color: transparentify($c-error, #fff, .25)
 
-  &.naked
-    transition: background-color color $ad.s
+  &.secondary
+    transition: 0.2s all ease-in-out
+    color: $tc-deep-gray
+    background-color: white
+    box-shadow: 0px 2px 2px 0px rgba(0 0 0 4.5%), inset 0px 0px 0px 1px white
+    background: linear-gradient(to bottom, white, white) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
     border: 1px solid transparent
-    background-color: transparent
-    font-weight: $fw.normal
+
+
+    &:hover
+      transition: 0.2s all ease-in-out
+      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
+
+    +focus-visible()
+      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
+
+    &:active
+      transition: 0.2s all ease-in-out
+      background: linear-gradient(to bottom, $c-focus, $c-focus) padding-box, linear-gradient(to bottom, #E8E7EC, #D7D7D7) border-box
+      box-shadow: none
+
+  &.naked
+    transition: 0.2s all ease-in-out
+    border: none
+    background-color: none
+    box-shadow: unset
     color: $tc-subtle-gray
 
     &.disabled
@@ -148,12 +194,19 @@
           color: hoverize($tc-error)
 
   &.only-icon
+    width: 2.5rem
+    padding: 0 $cs.xs 0 0.73rem
+
+  &.with-dropdown
+    padding: 0 $cs.xxs 0 $cs.xs
+
+  &.with-dropdown&.only-icon
+    width: fit-content
     .icon
-      margin-right: - $cs.xxs
+      margin-left: 0
 
   &.with-icon
-    .icon
-      margin-right: $cs.xxs
+    line-height: 1.3
 
     &.responsive-label
       padding: $cs.xs
@@ -181,17 +234,6 @@
 
     stop
       stop-color: $s
-
-@keyframes shake
-  0%
-    transform: translate(0)
-  25%
-    transform: translate(.4em)
-  75%
-    transform: translate(-.4em)
-  100%
-    transform: translate(0)
-
 .content
   position: relative
