+++ pkg/webui/components/button/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback, forwardRef, useMemo } from 'react'
+import React, { useCallback, forwardRef, useMemo, useState } from 'react'
 import classnames from 'classnames'
 import { useIntl } from 'react-intl'
 
 import Link from '@ttn-lw/components/link'
 import Spinner from '@ttn-lw/components/spinner'
 import Icon from '@ttn-lw/components/icon'
+import Dropdown from '@ttn-lw/components/dropdown'
 
 import Message from '@ttn-lw/lib/components/message'
 
@@ -39,10 +40,13 @@ const assembleClassnames = ({
   danger,
   warning,
   primary,
+  secondary,
   naked,
   unstyled,
+  grey,
   icon,
   busy,
+  dropdownItems,
   className,
   error,
   disabled,
@@ -54,17 +58,29 @@ const assembleClassnames = ({
     [style.danger]: danger,
     [style.warning]: warning,
     [style.primary]: primary,
+    [style.secondary]: secondary,
     [style.naked]: naked,
     [style.busy]: busy,
+    [style.grey]: grey,
     [style.withIcon]: icon !== undefined && message,
     [style.onlyIcon]: icon !== undefined && !message,
+    [style.withDropdown]: Boolean(dropdownItems),
     [style.error]: error && !busy,
     [style.disabled]: disabled || busy,
   })
 }
 
 const buttonChildren = props => {
-  const { icon, busy, message, children } = props
+  const {
+    dropdownItems,
+    icon,
+    busy,
+    message,
+    expanded,
+    noDropdownIcon,
+    dropdownClassName,
+    children,
+  } = props
 
   const content = Boolean(children) ? (
     children
@@ -72,6 +88,21 @@ const buttonChildren = props => {
     <>
       {icon ? <Icon className={style.icon} icon={icon} /> : null}
       {message ? <Message content={message} className={style.linkButtonMessage} /> : null}
+      {dropdownItems ? (
+        <>
+          {!noDropdownIcon && (
+            <Icon
+              className={classnames(style.arrowIcon, {
+                [style['arrow-icon-expanded']]: expanded,
+              })}
+              icon="expand_more"
+            />
+          )}
+          <Dropdown className={classnames(dropdownClassName)} open={expanded}>
+            {dropdownItems}
+          </Dropdown>
+        </>
+      ) : null}
     </>
   )
 
@@ -87,6 +118,7 @@ const Button = forwardRef((props, ref) => {
   const {
     autoFocus,
     disabled,
+    dropdownItems,
     name,
     type,
     value,
@@ -97,8 +129,27 @@ const Button = forwardRef((props, ref) => {
     form,
     ...rest
   } = props
+  const [expanded, setExpanded] = useState(false)
 
   const dataProps = useMemo(() => filterDataProps(rest), [rest])
+
+  const handleClickOutside = useCallback(
+    e => {
+      if (ref.current && !ref.current.contains(e.target)) {
+        setExpanded(false)
+      }
+    },
+    [ref],
+  )
+
+  const toggleDropdown = useCallback(() => {
+    setExpanded(oldExpanded => {
+      const newState = !oldExpanded
+      if (newState) document.addEventListener('mousedown', handleClickOutside)
+      else document.removeEventListener('mousedown', handleClickOutside)
+      return newState
+    })
+  }, [handleClickOutside])
 
   const handleClick = useCallback(
     evt => {
@@ -106,12 +157,16 @@ const Button = forwardRef((props, ref) => {
         return
       }
 
+      if (dropdownItems) {
+        toggleDropdown()
+        return
+      }
       // Passing a value to the onClick handler is useful for components that
       // are rendered multiple times, e.g. in a list. The value can be used to
       // identify the component that was clicked.
       onClick(evt, value)
     },
-    [busy, disabled, onClick, value],
+    [busy, disabled, dropdownItems, onClick, toggleDropdown, value],
   )
 
   const intl = useIntl()
@@ -127,7 +182,7 @@ const Button = forwardRef((props, ref) => {
     <button
       className={buttonClassNames}
       onClick={handleClick}
-      children={buttonChildren(props)}
+      children={buttonChildren({ ...props, expanded })}
       disabled={busy || disabled}
       ref={ref}
       {...htmlProps}
@@ -225,6 +280,8 @@ const commonPropTypes = {
   autoFocus: PropTypes.bool,
   /** A message to be evaluated and passed to the <button /> element. */
   title: PropTypes.message,
+  /** Dropdown items of the button. */
+  dropdownItems: PropTypes.node,
 }
 
 buttonChildren.propTypes = {
@@ -234,6 +291,7 @@ buttonChildren.propTypes = {
    */
   busy: commonPropTypes.busy,
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]),
+  expanded: PropTypes.bool,
   icon: commonPropTypes.icon,
   message: commonPropTypes.message,
 }
@@ -243,6 +301,7 @@ buttonChildren.defaultProps = {
   icon: undefined,
   message: undefined,
   children: null,
+  expanded: false,
 }
 
 Button.propTypes = {
@@ -253,6 +312,10 @@ Button.propTypes = {
    */
   onClick: PropTypes.func,
   ...commonPropTypes,
+}
+
+Button.defaultProps = {
+  onClick: () => null,
 }
 
 LinkButton.propTypes = {
