+++ pkg/webui/components/profile-dropdown-v2/index.js
@@ -18,7 +18,7 @@ import classnames from 'classnames'
 import Icon from '@ttn-lw/components/icon'
 import Dropdown from '@ttn-lw/components/dropdown-v2'
 import ProfilePicture from '@ttn-lw/components/profile-picture'
-import Button from '@ttn-lw/components/button-v2'
+import Button from '@ttn-lw/components/button'
 import style from '@ttn-lw/components/button-v2/button.styl'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
