+++ pkg/webui/components/sidebar/section-label/index.js
@@ -15,7 +15,7 @@
 import React from 'react'
 import classnames from 'classnames'
 
-import Button from '@ttn-lw/components/button-v2'
+import Button from '@ttn-lw/components/button'
 
 import Message from '@ttn-lw/lib/components/message'
